Pr3 - Add dtos to restcontrollers
PR

+++ src/main/java/com/example/nom035/repository/SurveyRepository.java
@@ -3,5 +3,5 @@ package com.example.nom035.repository;
 import com.example.nom035.entity.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyRepository extends JpaRepository<Survey, Integer> {
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
 }
