+++ src/main/java/com/example/nom035/service/CompanyService.java
@@ -18,7 +18,7 @@ public class CompanyService {
         return companyRepository.findAll();
     }
 
-    public Optional<Company> getCompanyById(Integer id) {
+    public Optional<Company> getCompanyById(Long id) {
         return companyRepository.findById(id);
     }
 
@@ -26,7 +26,7 @@ public class CompanyService {
         return companyRepository.save(company);
     }
 
-    public void deleteCompany(Integer id) {
+    public void deleteCompany(Long id) {
         companyRepository.deleteById(id);
     }
 }
